Generate data with population structure

This function allows you to simulate structured genetic data (SNP) with an unobserved environmental confounding effect.

Usage

genDataPS(
  n = 200,
  p = 1000,
  p1 = floor(p/2),
  nJ = rep(50, 4),
  structureX = "indep_subpops",
  Fst = NULL,
  inbr = "heterogeneous",
  structureGamma = c("dichotomous_discordant"),
  eta = 0.8,
  xi = 0,
  standardizeX = TRUE,
  structureX_other = NULL,
  sampleCols = TRUE,
  B0
)

Arguments

n

Number of observations/samples to simulate. Defaults to 200.

p

Number of SNPs to simulate. Defaults to 1000.

p1

Number of SNPs that are causal. Defaults to floor(p/2).

nJ

Number of observations in each subpopulation. The length of nJ corresponds to the number of subpopulations. Defaults to rep(50, 4).

structureX

Type of structure to simulate.

Fst

The desired final FST of the admixed individuals. Ranges from 0 to 1. A high FST implies greater differentiation among populations. Defaults to 0.1 if structureX = 1d_linear and 0.2 if structureX = indep_subpops. Otherwise defaults to NULL.

inbr

Indicates whether the desired inbreeding is homogeneous or heterogeneous. Defaults to heterogeneous.

structureGamma

The desired structure of the environmental confounding effect. Defaults to 'dichotomous_discordant'

eta

The desired proportion of variance in the outcome that is attributable to causal SNP effects. In other words, SNR.

xi

The desired proportion of the non-signal variance in the outcome that is attributable to unobserved environmental confounding effects.

standardizeX

Should the generated X matrix be standardized? Defaults to TRUE.

structureX_other

If structureX == "other", an matrix or SnpMatrix object with subjects in rows and SNPs in columns to be used to generate pseudophenotypes must be supplied here.

sampleCols

A logical flag for whether the columns of the resultant X matrix should be scrambled. This may be desirable if the causal SNPs should change from one simulation to the next. Defaults to TRUE.

B0

Optional. Additional intercept value.

Value

A list with 7 elements: a matrix of SNP data (X), a single-column matrix of outcome values (y), a single column matrix 'env', a vector of coefficients (beta), a matrix of Z values allocating environmental effects among subjects (Z), a vector of numeric values representing environmental effects (gamma), and the type of structure used in the SNP data (structureX)

Examples

sim_dat <- genDataPS(structureX = "1d_linear")
example_fit <- plmm(sim_dat$X, sim_dat$y, V = sim_dat$X%*%t(sim_dat$X))