Predict method for a list used in cross-validation (e.g. the fit.i within cvf)

Predict method for a list used in cross-validation (e.g. the fit.i within cvf)

Usage

# S3 method for list
predict(
  fit,
  newX,
  type = c("lp", "coefficients", "vars", "nvars", "blup"),
  lambda,
  idx = 1:length(fit$lambda),
  prep = NULL,
  V11 = NULL,
  V21 = NULL,
  ...
)

Arguments

fit

A list with the components returned by plmm_fit

newX

A design matrix used for computing predicted values if requested.

type

A character argument indicating what type of prediction should be returned. Options are "lp," "coefficients," "vars," "nvars," and "blup." See details.

lambda

A numeric vector of regularization parameter lambda values at which predictions are requested.

idx

Vector of indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned.

prep

Optional argument. Result of the call to plmm_prep which corresponds to the fit argument. Required if type == 'blup' and object is too large to be returned in fit object.

V11

Variance-covariance matrix of the training data. Extracted from estimated_V that is generated using all observations.

V21

Covariance matrix between the training and the testing data. Extracted from estimated_V that is generated using all observations.

...

Additional optional arguments

covariance

Optional argument. $q times n$ covariance matrix between new and old observations. Required if type == 'blup'.

Details

Define beta-hat as the coefficients estimated at the value of lambda that minimizes cross-validation error (CVE). Then options for type are as follows:

'response' (default): uses the product of newX and beta-hat to predict new values of the outcome. This does not incorporate the correlation structure of the data. For the stats folks out there, this is simply the linear predictor.

'blup' (acronym for Best Linear Unbiased Predictor): adds to the 'response' a value that represents the esetimated random effect. This addition is a way of incorporating the estimated correlation structure of data into our prediction of the outcome.

'coefficients': returns the estimated beta-hat

'vars': returns the indicies of variables (e.g., SNPs) with nonzero coefficients at each value of lambda. EXCLUDES intercept.

'nvars': returns the number of variables (e.g., SNPs) with nonzero coefficients at each value of lambda. EXCLUDES intercept.