Calculate scale by the population standard deviation, without centering

This function allows you to scale vectors of a matrix by their population standard deviation without centering; we assume our sample is the population.

Usage

scale_varp(X)

Arguments

X

numeric matrix

Examples

M <- matrix(rnorm(25), 5, 5)
head(M)
#>            [,1]       [,2]        [,3]        [,4]       [,5]
#> [1,] -0.1796834 -0.3973511 -1.08182141  2.07028732 -0.9554772
#> [2,]  0.2816343 -0.7614084 -0.04093868 -0.02847675  0.4800143
#> [3,] -2.4594117  0.5514024 -0.16712948  2.76619169  1.5700993
#> [4,]  1.3998633  0.6812048 -0.07798380  0.16846559 -0.1422800
#> [5,] -1.0880932  1.1009150 -1.43406534  1.69374683  0.2957709
M_scaled <- scale_varp(M)
head(M_scaled)
#>            [,1]       [,2]        [,3]        [,4]       [,5]
#> [1,] -0.1317228 -0.5389597 -1.33924678  1.20187424 -1.1082832
#> [2,]  0.2064613 -1.0327605 -0.05068026 -0.01653175  0.5567812
#> [3,] -1.8029527  0.7479121 -0.20689886  1.60587108  1.8211996
#> [4,]  1.0262159  0.9239738 -0.09654048  0.09780017 -0.1650343
#> [5,] -0.7976625  1.4932609 -1.77530909  0.98327931  0.3430724

X_scaled <- scale_varp(admix$X)
admix$X[1:5, 1:7]; X_scaled[1:5, 1:7]
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1
#>          Snp1 Snp2 Snp3     Snp4 Snp5    Snp6     Snp7
#> [1,] 0.000000    0    0 0.000000    0 0.00000 1.599513
#> [2,] 0.000000    0    0 0.000000    0 1.24059 0.000000
#> [3,] 0.000000    0    0 0.000000    0 1.24059 1.599513
#> [4,] 0.000000    0    0 0.000000    0 0.00000 0.000000
#> [5,] 2.276886    0    0 1.522382    0 0.00000 1.599513