Predict method for plmm class

Predict method for plmm class

Usage

# S3 method for plmm
predict(
  object,
  newX,
  type = c("response", "coefficients", "vars", "nvars", "blup"),
  lambda,
  idx = 1:length(object$lambda),
  X,
  y,
  ...
)

Arguments

object

An object of class plmm.

newX

Design matrix used for computing predicted values if requested.

type

A character argument indicating what type of prediction should be returned. Options are "response," "coefficients," "vars," "nvars," and "blup." See details.

lambda

A numeric vector of regularization parameter lambda values at which predictions are requested.

idx

Vector of indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned.

X

Optional argument. Original design matrix (not including intercept column) from object. Required if type == 'blup' and object is too large to be returned in plmm object.

y

Optional argument. Original continuous outcome vector from object. Required if type == 'blup'.

...

Additional optional arguments

U

Optional argument. Eigenvectors from the similarity matrix from object. Required if type == 'blup'.

S

Optional argument. Eigenvalues from the similarity matrix from object. Required if type == 'blup'.

eta

Optional argument. Estimated $eta$ value from object. Required if type == 'blup'.

covariance

Optional argument. $q times n$ covariance matrix between new and old observations. Required if type == 'blup'.

Examples

# fit a model 
fit <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))

# simulate new data 
newX <- sim_ps_x(n = nrow(admix$X), nJ = 4, p = ncol(admix$X),
 structureX = "independent", inbr = "heterogeneous", standardizeX = FALSE)
 
 # make predictions for all lambda values 
 pred1 <- predict(object = fit, newX = newX, type = "response")
 
 # make predictions for a select number of lambda values 
 pred2 <- predict(object = fit, newX = newX, type = "response", idx=98)
 
 # make prediction using blup 
 pred3 <- predict(object = fit, newX = newX, type = "blup", idx=98)
#> Warning: The BLUP option is under development. Rely on these estimates at your own risk.
#> Error in predict.plmm(object = fit, newX = newX, type = "blup", idx = 98): The design matrix is required for BLUP calculation, but is not available in the plmm object.
#>  This is ususally because X is large.
#>  Please supply the no-intercept design matrix to the X argument, and the vector of outcomes to the y argument.

 # compare y predictions 
  compare_y <- data.frame(y = admix$y, yhat_response = pred2, yhat_blup = pred3)
#> Error in data.frame(y = admix$y, yhat_response = pred2, yhat_blup = pred3): object 'pred3' not found
 
 if (FALSE) {
 fit_noX <- plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X),
  returnX = FALSE)
 
 pred4 <- predict(object = fit_noX, newX = newX, type = "blup", idx=98)
 
 }
 
 
# make predictions when X is big
cad_mid <- process_plink(prefix = "cad_mid", dataDir = plink_example(path="cad_mid.fam", parent=T))
#> 
#> Preprocessing cad_mid data:
#> Error in base::system.file("inst/extdata", package = "penalizedLMM", mustWork = T): no file found
cad_clinical <- read.csv(plink_example(path="cad_clinical.csv"))
#> Error in base::system.file("inst/extdata", path, package = "penalizedLMM",     mustWork = TRUE): no file found
# for the sake of illustration, I use a simple mean imputation for the outcome 
cad_clinical$hdl_impute <- ifelse(is.na(cad_clinical$hdl), mean(cad_clinical$hdl, na.rm = T), cad_clinical$hdl)
#> Error in ifelse(is.na(cad_clinical$hdl), mean(cad_clinical$hdl, na.rm = T),     cad_clinical$hdl): object 'cad_clinical' not found
fit_cad <- plmm(X = cad_mid$genotypes, y = cad_clinical$hdl_impute, k = 5)
#> Error in "SnpMatrix" %in% class(X): object 'cad_mid' not found
cad_X <- cad_mid$genotypes
#> Error in eval(expr, envir, enclos): object 'cad_mid' not found
cad_y <- cad_clinical$hdl_impute
#> Error in eval(expr, envir, enclos): object 'cad_clinical' not found
newX_cad <- sim_ps_x(n = nrow(cad_X), nJ = 4, p = ncol(cad_X),
 structureX = "independent", inbr = "heterogeneous", standardizeX = FALSE)
#> Error in nrow(cad_X): object 'cad_X' not found
pred_cad <- predict(object = fit_cad, newX = newX_cad, type='blup', idx = 95, X = cad_X, y = cad_y)
#> Error in predict(object = fit_cad, newX = newX_cad, type = "blup", idx = 95,     X = cad_X, y = cad_y): object 'fit_cad' not found
head(data.frame(cad_y, pred_cad))
#> Error in data.frame(cad_y, pred_cad): object 'cad_y' not found