Calculate scale by the population standard deviation, without centering

This function allows you to scale vectors of a matrix by their population standard deviation without centering; we assume our sample is the population.

Usage

scale_varp(X)

Arguments

X

numeric matrix

Examples

M <- matrix(rnorm(25), 5, 5)
head(M)
#>             [,1]       [,2]       [,3]        [,4]       [,5]
#> [1,]  0.68944281  1.1972054  0.3049187  0.06913725  0.2474286
#> [2,]  0.09086577 -1.4755092 -0.2495615 -0.97412774  1.5351675
#> [3,]  0.45291872  0.7995004 -0.8053116  0.86949771  0.3353290
#> [4,] -0.83686321  1.3364604 -0.9206062 -0.03943486  0.2426246
#> [5,] -0.46284413 -2.9255514  0.4117681 -1.80970428 -0.9578383
M_scaled <- scale_varp(M)
head(M_scaled)
#> $scale_vals
#> [1] 0.5662664 1.7084851 0.6034653 0.9986306 0.8374636
#> 
#> $scaled_X
#>            [,1]       [,2]       [,3]        [,4]       [,5]
#> [1,]  1.2175238  0.7007409  0.5052796  0.06923205  0.2954500
#> [2,]  0.1604647 -0.8636360 -0.4135473 -0.97546350  1.8331154
#> [3,]  0.7998333  0.4679587 -1.3344787  0.87069000  0.4004103
#> [4,] -1.4778614  0.7822488 -1.5255329 -0.03948893  0.2897136
#> [5,] -0.8173611 -1.7123658  0.6823393 -1.81218582 -1.1437372
#> attr(,"scaled:scale")
#> [1] 0.5662664 1.7084851 0.6034653 0.9986306 0.8374636
#> 

X_scaled <- scale_varp(admix$X)
admix$X[1:5, 1:7]; X_scaled[1:5, 1:7]
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1
#> Error in X_scaled[1:5, 1:7]: incorrect number of dimensions
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1