Calculate scale by the population standard deviation, without centering

This function allows you to scale vectors of a matrix by their population standard deviation without centering; we assume our sample is the population.

Usage

scale_varp(X)

Arguments

X

numeric matrix

Examples

M <- matrix(rnorm(25), 5, 5)
head(M)
#>            [,1]        [,2]        [,3]        [,4]       [,5]
#> [1,]  0.5716664 -0.21181425  0.05005486 -0.37561533 -0.2358852
#> [2,]  0.5982258  2.16116286  0.35529428  0.99662047  2.0718074
#> [3,] -0.2866903 -0.80159299  0.29430145 -0.04274144  1.4213490
#> [4,] -0.2299971 -0.80755942 -1.69391274 -0.60978010 -0.5394943
#> [5,] -1.3641039 -0.09054345 -0.22573025 -1.52705909 -1.2146765
M_scaled <- scale_varp(M)
head(M_scaled)
#> $scale_vals
#> [1] 0.7321948 1.0971228 0.7919151 0.8763455 1.2755198
#> 
#> $scaled_X
#>            [,1]       [,2]        [,3]        [,4]       [,5]
#> [1,]  0.7807573 -0.1930634  0.06320736 -0.42861557 -0.1849326
#> [2,]  0.8170309  1.9698460  0.44865200  1.13724605  1.6242848
#> [3,] -0.3915492 -0.7306320  0.37163259 -0.04877236  1.1143292
#> [4,] -0.3141201 -0.7360702 -2.13900809 -0.69582156 -0.4229604
#> [5,] -1.8630342 -0.0825281 -0.28504351 -1.74253084 -0.9522992
#> attr(,"scaled:scale")
#> [1] 0.7321948 1.0971228 0.7919151 0.8763455 1.2755198
#> 

X_scaled <- scale_varp(admix$X)
admix$X[1:5, 1:7]; X_scaled[1:5, 1:7]
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1
#> Error in X_scaled[1:5, 1:7]: incorrect number of dimensions
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1