Calculate scale by the population standard deviation, without centering

This function allows you to scale vectors of a matrix by their population standard deviation without centering; we assume our sample is the population.

Usage

scale_varp(X)

Arguments

X

numeric matrix

Examples

M <- matrix(rnorm(25), 5, 5)
head(M)
#>             [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  1.51508103 -1.0294957 -1.4502181  0.3361512 0.30860017
#> [2,]  0.43425977 -1.5602073  0.1139973 -2.0317516 0.53233981
#> [3,] -0.09187143 -1.7451543 -0.6480774 -1.8902529 0.64230000
#> [4,] -1.40422349 -0.5870656  0.4467984  1.5540705 0.03654447
#> [5,]  0.70411914 -0.3513557  1.0093184 -1.5238120 0.26566515
M_scaled <- scale_varp(M)
head(M_scaled)
#> $scale_vals
#> [1] 0.9960040 1.1838723 0.8665431 1.5843741 0.4154714
#> 
#> $scaled_X
#>             [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  1.52115958 -0.8696003 -1.6735672  0.2121666 0.74277112
#> [2,]  0.43600203 -1.3178847  0.1315541 -1.2823686 1.28129105
#> [3,] -0.09224002 -1.4741068 -0.7478883 -1.1930597 1.54595472
#> [4,] -1.40985729 -0.4958859  0.5156101  0.9808735 0.08795906
#> [5,]  0.70694409 -0.2967852  1.1647642 -0.9617754 0.63943062
#> attr(,"scaled:scale")
#> [1] 0.9960040 1.1838723 0.8665431 1.5843741 0.4154714
#> 

X_scaled <- scale_varp(admix$X)
admix$X[1:5, 1:7]; X_scaled[1:5, 1:7]
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1
#> Error in X_scaled[1:5, 1:7]: incorrect number of dimensions
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1