Calculate scale by the population standard deviation, without centering

This function allows you to scale vectors of a matrix by their population standard deviation without centering; we assume our sample is the population.

Usage

scale_varp(X)

Arguments

X

numeric matrix

Examples

M <- matrix(rnorm(25), 5, 5)
head(M)
#>             [,1]       [,2]        [,3]       [,4]        [,5]
#> [1,] -1.07930890  1.4097644  0.63716904 -0.2271730 -0.42695472
#> [2,] -0.56287819  1.0372643 -0.05925889 -0.2468695 -0.20413139
#> [3,]  0.02059107 -0.2506698  1.43191042  0.8928686  0.45558410
#> [4,] -0.53573007 -2.1978570 -0.25819385  0.6386396  0.02040667
#> [5,]  0.71168978 -0.5731428 -1.42227361  1.2981014  0.90887403
M_scaled <- scale_varp(M)
head(M_scaled)
#>             [,1]       [,2]        [,3]       [,4]        [,5]
#> [1,] -1.59983749  1.0951625  0.66797944 -0.2931475 -0.85119497
#> [2,] -0.83434282  0.8057892 -0.06212436 -0.3185642 -0.40696495
#> [3,]  0.03052172 -0.1947305  1.50115066  1.1521711  0.90827171
#> [4,] -0.79410172 -1.7073850 -0.27067885  0.8241102  0.04068359
#> [5,]  1.05492319 -0.4452407 -1.49104785  1.6750897  1.81196964

X_scaled <- scale_varp(admix$X)
admix$X[1:5, 1:7]; X_scaled[1:5, 1:7]
#>      Snp1 Snp2 Snp3 Snp4 Snp5 Snp6 Snp7
#> [1,]    0    0    0    0    0    0    1
#> [2,]    0    0    0    0    0    1    0
#> [3,]    0    0    0    0    0    1    1
#> [4,]    0    0    0    0    0    0    0
#> [5,]    1    0    0    1    0    0    1
#>          Snp1 Snp2 Snp3     Snp4 Snp5    Snp6     Snp7
#> [1,] 0.000000    0    0 0.000000    0 0.00000 1.599513
#> [2,] 0.000000    0    0 0.000000    0 1.24059 0.000000
#> [3,] 0.000000    0    0 0.000000    0 1.24059 1.599513
#> [4,] 0.000000    0    0 0.000000    0 0.00000 0.000000
#> [5,] 2.276886    0    0 1.522382    0 0.00000 1.599513